Moved NN implementation to own files and added import safeguard
