+++ cosipy/response/__init__.py
@@ -9,5 +9,3 @@ from .threeml_point_source_response import *
 from .threeml_extended_source_response import *
 from .instrument_response import *
 from .rsp_to_arf_rmf import RspArfRmfConverter
-from .NNResponse import NNResponse
-from .instrument_response_function import *
